remove some code duplication in fragment classes
